refactor: extracted more logic into hooks and function
