+++ src/components/Chat/index.tsx
@@ -1,11 +1,9 @@
 'use client'
 
-import { SignOutButton } from '../buttons'
+import { SignOutButton } from '../AuthButtons'
 import { useSession } from 'next-auth/react'
 
-interface ChatProps {}
-
-export const Chat: React.FC<ChatProps> = () => {
+export const Chat: React.FC = () => {
 	const { data: session } = useSession()
 	return (
 		<div className='m-auto flex items-center gap-3 flex-col justify-center h-screen'>
