+++ src/components/Message/index.tsx
@@ -1,19 +1,11 @@
 import { Message as MessageType } from '@/graphql/types'
 import { useAutoScroll } from '@/hooks/useAutoScroll'
-import { formatTime, is } from '@/util/functions'
-import { Skeleton, UserIcon } from '@components'
-import { useSession } from 'next-auth/react'
-import { useContext } from 'react'
-import { ConversationContext } from '../Context/ConversationContext'
 import useSubscribeToMessages from '@/hooks/useSubscribeToMessages'
+import { formatTime, is } from '@/util/functions'
+import { MessageProps, Skeleton, UserIcon } from '@components'
+import { useSession } from 'next-auth/react'
 
-const testCases = [
-	'Lorem ipsum dolor sit amet, consectetur adi',
-	'LoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLor eLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLoreLore',
-	'Lorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adiLorem ipsum dolor sit amet, consectetur adi'
-]
-
-export const Message = ({ data }: { data: MessageType }) => {
+export const Message = ({ data }: MessageProps) => {
 	const { data: userData } = useSession()
 	const { body, createdAt, sender } = data
 	const right = sender.id === userData?.user.id
@@ -24,14 +16,6 @@ export const Message = ({ data }: { data: MessageType }) => {
 				<p className='text-sm'>
 					{sender.username} <span className='text-zinc-400 text-sm'>{formatTime(createdAt)}</span>
 				</p>
-				{/* {testCases.map((e, index) => (
-					<div
-						key={index}
-						className={`xl:max-w-[60%] lg:max-w-[75%] max-w-[calc(100%-48px)] w-fit break-words whitespace-pre-wrap p-2 rounded ${is(right, 'bg-sky-900', 'bg-zinc-700')} text-sm`}
-					>
-						{e}
-					</div>
-				))} */}
 				<div className={`xl:max-w-[60%] lg:max-w-[75%] max-w-[calc(100%-48px)] w-fit break-words whitespace-pre-wrap p-2 rounded ${is(right, 'bg-sky-900', 'bg-zinc-700')} text-sm`}>
 					{body}
 				</div>
