+++ src/components/UserIcon/index.tsx
@@ -1,8 +1,9 @@
 import DefaultUserIcon from '@assets/default-user.png'
 import Image from 'next/image'
 import { is } from '@/util/functions'
+import { UserIconProps } from '@components'
 
-export const UserIcon = ({ right = false, notification }: { right?: boolean; notification?: boolean }) => {
+export const UserIcon = ({ right = false, notification }: UserIconProps) => {
 	return (
 		<div className={`min-h-[40px] min-w-[40px] relative ${is(right, 'order-last')}`}>
 			<Image src={DefaultUserIcon} alt={'user icon'} className={`h-10 w-10 ${is(right, 'order-last')}`} />
