+++ src/components/index.ts
@@ -1,5 +1,5 @@
 export * from './Auth'
-export * from './buttons'
+export * from './AuthButtons'
 export * from './Chat'
 export * from './ChatList'
 export * from './Conversation'
@@ -13,6 +13,7 @@ export * from './Skeleton'
 export * from './StartConversationButton'
 export * from './StartConversationForm'
 export * from './svgComponents'
+export * from './types'
 export * from './UserIcon'
 export * from './UserList'
 export * from './UserTags'
