+++ src/context/ConversationContext.tsx
@@ -1,20 +1,14 @@
 import MessageOperations from '@/graphql/operations/message'
-import { Message, MutationSendMessageArgs, QueryMessagesArgs } from '@/graphql/types'
+import { MutationSendMessageArgs } from '@/graphql/types'
+import addMessageToCache from '@/graphql/util/addMessageToCache'
 import { useAutoResize } from '@/hooks/useAutoResize'
 import { OpReturnType } from '@/util/utilityTypes'
-import { useMutation, useQuery } from '@apollo/client'
+import { useMutation } from '@apollo/client'
 import ObjectID from 'bson-objectid'
 import { useSession } from 'next-auth/react'
-import { Dispatch, KeyboardEvent, ReactNode, RefObject, SetStateAction, createContext, useContext, useEffect, useState } from 'react'
+import { KeyboardEvent, createContext, useContext, useState } from 'react'
 import { AppContext } from './AppContext'
-
-interface ConversationContextValues {
-	message: string
-	setMessage: Dispatch<SetStateAction<string>>
-	onSendMessage: () => Promise<void>
-	onKeyDown: <T extends HTMLElement>(event: KeyboardEvent<T>) => Promise<void>
-	elementRef: RefObject<HTMLTextAreaElement> | null
-}
+import { ContextProps, ConversationContextValues } from './types'
 
 const defaultContext: ConversationContextValues = {
 	message: '',
@@ -26,13 +20,7 @@ const defaultContext: ConversationContextValues = {
 
 export const ConversationContext = createContext<ConversationContextValues>(defaultContext)
 
-interface ConversationContextProps {
-	children: ReactNode
-}
-
-type MessageWithSenderId = Message & { senderId: string }
-
-export const ConversationContextProvider: React.FC<ConversationContextProps> = ({ children }) => {
+export const ConversationContextProvider: React.FC<ContextProps> = ({ children }) => {
 	const { selectedConversation } = useContext(AppContext)
 	const [message, setMessage] = useState('')
 	const { data: userData } = useSession()
@@ -56,36 +44,7 @@ export const ConversationContextProvider: React.FC<ConversationContextProps> = (
 			const { data, errors } = await sendMessage({
 				variables: newMessage,
 				optimisticResponse: { sendMessage: new ObjectID().toString() },
-				update: (cache, data) => {
-					if (!data.data) return
-					const existing = cache.readQuery<{ messages: MessageWithSenderId[] }>({
-						query: MessageOperations.Query.messages,
-						variables: { conversationId }
-					})
-					const { email, id, username, image } = userData.user
-					if (!existing) return
-					cache.writeQuery<{ messages: MessageWithSenderId[] }, { conversationId: string }>({
-						query: MessageOperations.Query.messages,
-						variables: { conversationId },
-						data: {
-							...existing,
-							messages: [
-								...existing.messages,
-								{
-									...newMessage,
-									id: data.data?.sendMessage as string,
-									createdAt: new Date(Date.now()),
-									sender: {
-										id,
-										username,
-										email: email as string,
-										image
-									}
-								}
-							]
-						}
-					})
-				}
+				update: (cache, data) => addMessageToCache({ newMessage, cache, data, userData })
 			})
 			if (!data?.sendMessage || errors) throw new Error('Failed to send message')
 		} catch (error) {
