+++ src/graphql/middleware/authentication.ts
@@ -1,11 +1,10 @@
 import { setContext } from '@apollo/client/link/context'
 import { getSession } from 'next-auth/react'
 
-export const authMiddleware = setContext(async (operation, { headers }) => {
+export const authMiddleware = setContext(async (_, { headers }) => {
 	try {
 		const session = await getSession()
 
-		// const { user } = await fetch('/api/session').then((res) => res.json())
 		return {
 			headers: {
 				...headers,
